#19 feat : RegisterPage 불필요한 부분 삭제
